feat: groceryList belongsTo User

+++ app/Http/Controllers/GroceryListController.php
@@ -8,34 +8,41 @@ use App\Http\Resources\GroceryListResource;
 use App\Models\GroceryList;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class GroceryListController extends Controller
 {
     public function index(): AnonymousResourceCollection
     {
         $groceryLists = GroceryList::query()
-            ->with([
-                'recipes:id,name,picture,preparation_time,total_time,kcal'
-            ])
+            ->forMe()
+            ->with(['recipes:id,name,picture,preparation_time,total_time,kcal'])
             ->latest()
             ->paginate();
 
         return GroceryListResource::collection($groceryLists);
     }
 
-    public function show(GroceryList $groceryList): GroceryListResource
-    {
-        return GroceryListResource::make($groceryList->load('recipes.ingredients'));
-    }
-
     public function last(): GroceryListResource
     {
-        return GroceryListResource::make(GroceryList::query()->latest()->first());
+        return GroceryListResource::make(
+            GroceryList::query()
+                ->forMe()
+                ->latest()
+                ->firstOrFail()
+        );
+    }
+
+    public function show(GroceryList $groceryList): GroceryListResource
+    {
+        Gate::authorize('view', $groceryList);
+
+        return GroceryListResource::make($groceryList->load('recipes.ingredients'));
     }
 
     public function store(GroceryListStoreRequest $request): GroceryListResource
     {
-        $groceryList = GroceryList::query()->create($request->getInput());
+        $groceryList = $request->user()->groceryLists()->create($request->getInput());
 
         $groceryList->recipes()->sync($request->recipes);
 
@@ -44,9 +51,11 @@ class GroceryListController extends Controller
 
     public function update(GroceryListUpdateRequest $request, GroceryList $groceryList): GroceryListResource
     {
+        Gate::authorize('update', $groceryList);
+
         $groceryList->update($request->getInput());
 
-        if($request->recipes) {
+        if ($request->recipes) {
             $groceryList->recipes()->sync($request->recipes);
         }
 
@@ -55,6 +64,8 @@ class GroceryListController extends Controller
 
     public function destroy(GroceryList $groceryList): Response
     {
+        Gate::authorize('delete', $groceryList);
+
         $groceryList->delete();
 
         return response()->noContent();
